Dev Mode
Add a new tab for testing that finds the difference from a reference save and the editor's current export. Useful for checking if you missed something in a new version, or otherwise messed up something. I'll try not to include the dev tab on the public version.

Also some minor code changes and cleanup.